requirements.txt
pandas
python-binance==1.0.17
ta
urllib3==1.26.16
requests==2.28.2
aiocron
Pillow